requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
autopep8==2.3.2
beautifulsoup4==4.14.3
blis==1.3.3
catalogue==2.0.10
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
cloudpathlib==0.23.0
cltk==2.5.1
colorama==0.4.6
confection==1.3.3
cuda-bindings==13.2.0
cuda-pathfinder==1.5.3
cuda-toolkit==13.0.2
cymem==2.0.13
emoji==2.15.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.29.0
fsspec==2026.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.15
networkx==3.6.1
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
packaging==26.1
platformdirs==4.9.6
preshed==3.0.13
protobuf==7.34.1
pycodestyle==2.14.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
python-dotenv==1.2.2
python-slugify==8.0.4
python_speech_features==0.6
regex==2026.4.4
requests==2.33.1
rich==15.0.0
scikit-learn==1.8.0
scipy==1.17.1
selectolax==0.3.34
setuptools==81.0.0
shellingham==1.5.4
sltk==0.5
smart_open==7.6.0
soupsieve==2.8.3
spacy==3.8.14
spacy-html-tokenizer==0.1.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
stanza==1.11.1
sympy==1.14.0
termcolor==3.3.0
text-unidecode==1.3
thinc==8.3.13
threadpoolctl==3.6.0
torch==2.11.0
tqdm==4.67.3
triton==3.6.0
typer==0.24.2
typing-inspection==0.4.2
typing_extensions==4.15.0
udapi==0.5.2
udtools==0.2.7
urllib3==2.6.3
wasabi==1.1.3
weasel==1.0.0
wheel==0.47.0
wrapt==2.1.2
